+++ src/repos/tours/index.ts
@@ -1,6 +1,6 @@
 import { ExtensionContext } from "vscode";
 import { onDidEndTour } from "../../tour";
-import { RepoFileSystemProvider } from "../fileSystem";
+import { REPO_SCHEME } from "../fileSystem";
 import { store } from "../store";
 import { registerTourCommands } from "./commands";
 
@@ -8,7 +8,7 @@ export async function registerTourController(context: ExtensionContext) {
   registerTourCommands(context);
 
   onDidEndTour((tour) => {
-    if (tour.id.startsWith(`${RepoFileSystemProvider.SCHEME}:/`)) {
+    if (tour.id.startsWith(`${REPO_SCHEME}:/`)) {
       store.isInCodeTour = false;
     }
   });