+++ src/repos/wiki/commands.ts
@@ -3,7 +3,8 @@ import { EXTENSION_NAME } from "../../constants";
 import { stringToByteArray, withProgress } from "../../utils";
 import { RepoFileSystemProvider } from "../fileSystem";
 import { RepositoryFileNode, RepositoryNode } from "../tree/nodes";
-import { openRepoDocument, sanitizeName } from "../utils";
+import { openRepoDocument } from "../utils";
+import { getPageFilePath } from "./utils";
 
 import moment = require("moment");
 const { titleCase } = require("title-case");
@@ -17,25 +18,6 @@ function createWikiPage(name: string, repo: string, filePath: string) {
   return workspace.fs.writeFile(uri, stringToByteArray(fileHeading));
 }
 
-function getDailyDirectory() {
-  return workspace.getConfiguration("gistpad").get("wikis.dailyDirectory");
-}
-
-const DAILY_PATTERN = /\d{4}-\d{2}-\d{2}/;
-function getPageFile(name: string) {
-  let fileName = sanitizeName(name).toLocaleLowerCase();
-  if (!fileName.endsWith(".md")) {
-    fileName += ".md";
-  }
-
-  if (DAILY_PATTERN.test(fileName)) {
-    const dailyDirectory = getDailyDirectory();
-    return `${dailyDirectory}/${fileName}`;
-  } else {
-    return fileName;
-  }
-}
-
 export function registerCommands(context: ExtensionContext) {
   // This is a private command that handles dynamically
   // creating wiki documents, when the user auto-completes
@@ -44,7 +26,7 @@ export function registerCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}._createWikiPage`,
       async (repo: string, name: string) => {
-        const fileName = getPageFile(name);
+        const fileName = getPageFilePath(name);
         await createWikiPage(name, repo, fileName);
 
         // Automatically save the current, in order to ensure
@@ -63,7 +45,7 @@ export function registerCommands(context: ExtensionContext) {
         });
 
         if (page) {
-          const path = getPageFile(page);
+          const path = getPageFilePath(page);
           const filePath =
             node instanceof RepositoryFileNode
               ? `${node.file.path}/${path}`
@@ -83,15 +65,22 @@ export function registerCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.openTodayPage`,
       async (node: RepositoryNode) => {
-        const page = moment().format("YYYY-MM-DD");
-        const filePath = getPageFile(page);
+        const sharedMoment = moment();
+        const fileName = sharedMoment.format("YYYY-MM-DD");
+        const filePath = getPageFilePath(fileName);
+
+        const titleFormat = workspace
+          .getConfiguration(EXTENSION_NAME)
+          .get("wikis.daily.titleFormat", "LL");
+
+        const pageTitle = sharedMoment.format(titleFormat);
 
         const uri = RepoFileSystemProvider.getFileUri(node.repo.name, filePath);
         const [, file] = RepoFileSystemProvider.getRepoInfo(uri)!;
 
         if (!file) {
           await withProgress("Adding new page...", async () => {
-            return createWikiPage(page, node.repo.name, filePath);
+            return createWikiPage(pageTitle, node.repo.name, filePath);
           });
         }